+++ fastlane_bot/events/exchanges/__init__.py
@@ -16,6 +16,7 @@ from fastlane_bot.events.exchanges.uniswap_v2 import UniswapV2
 from fastlane_bot.events.exchanges.uniswap_v3 import UniswapV3
 from fastlane_bot.events.exchanges.bancor_v2 import BancorV2
 from fastlane_bot.events.exchanges.bancor_pol import BancorPol
+from fastlane_bot.events.exchanges.balancer import Balancer
 
 # Create a single instance of ExchangeFactory
 exchange_factory = ExchangeFactory()
@@ -30,4 +31,4 @@ exchange_factory.register_exchange("bancor_v2", BancorV2)
 exchange_factory.register_exchange("bancor_pol", BancorPol)
 exchange_factory.register_exchange("pancakeswap_v2", PancakeswapV2)
 exchange_factory.register_exchange("pancakeswap_v3", PancakeswapV3)
-
+exchange_factory.register_exchange("balancer", Balancer)